add ld to lightdash pattern and move user attribute logic
